Fix secret shortening in oath app
The oath app attempts to shorten secrets (by hashing them) if they
exceed a specific length.

This change adjusts the calculation so that it uses the block size of
the hash algorithm to decide when to shorten the secret.

Previously, the output size was used, instead of the block size. This
meant that secrets that were added to the device that exceeded the
output size, but not the block size, would end up getting hashed twice.